refactor: :recycle: move `check_properties.py` into root code folder
