+++ src/seedcase_sprout/core/check_properties.py
