+++ src/seedcase_sprout/core/create_resource_properties.py
@@ -1,13 +1,13 @@
 from pathlib import Path
 
 from seedcase_sprout.core.check_is_dir import check_is_dir
+from seedcase_sprout.core.check_properties import (
+    check_resource_properties,
+)
 from seedcase_sprout.core.create_relative_resource_data_path import (
     create_relative_resource_data_path,
 )
 from seedcase_sprout.core.properties import ResourceProperties
-from seedcase_sprout.core.sprout_checks.check_properties import (
-    check_resource_properties,
-)
 
 
 def create_resource_properties(
