+++ src/seedcase_sprout/core/read_properties.py
@@ -1,9 +1,9 @@
 from pathlib import Path
 
 from seedcase_sprout.core.check_is_file import check_is_file
+from seedcase_sprout.core.check_properties import check_properties
 from seedcase_sprout.core.properties import PackageProperties
 from seedcase_sprout.core.read_json import read_json
-from seedcase_sprout.core.sprout_checks.check_properties import check_properties
 
 
 def read_properties(path: Path) -> PackageProperties:
