+++ tests/core/test_examples.py
@@ -1,3 +1,7 @@
+from seedcase_sprout.core.check_properties import (
+    check_package_properties,
+    check_resource_properties,
+)
 from seedcase_sprout.core.examples import (
     example_data,
     example_data_all_types,
@@ -6,10 +10,6 @@ from seedcase_sprout.core.examples import (
     example_resource_properties_all_types,
 )
 from seedcase_sprout.core.sprout_checks.check_data import check_data
-from seedcase_sprout.core.sprout_checks.check_properties import (
-    check_package_properties,
-    check_resource_properties,
-)
 
 
 def test_creates_correct_package_properties_object():
